backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
supabase==2.7.4
httpx==0.27.2
python-telegram-bot==21.5
pydantic==2.9.2
africastalking==1.2.7
resend==2.4.0
